+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 - Improved keyboard and screen reader experience for `EuiColorPicker` ([#4886](https://github.com/elastic/eui/pull/4886))
 - Changed `EuiImage` to use `ImgHTMLAttributes` type ([#4865](https://github.com/elastic/eui/pull/4865))
+- Added `current` as a `status` of `EuiHorizontalStep` ([#4911](https://github.com/elastic/eui/pull/4911))
 
 **Bug fixes**
 
 - Fixed focus bug in `EuiColorPicker` which allowed user to break out of focus lock ([#4886](https://github.com/elastic/eui/pull/4886))
-- Fixed default status of `EuiHorizontalSteps` to be undefined ([#4911](https://github.com/elastic/eui/pull/4911))
 
 ## [`35.0.0`](https://github.com/elastic/eui/tree/v35.0.0)